backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
cryptography==42.0.0
pyotp==2.9.0

# Communication
aiosmtplib==3.0.1
twilio==8.11.1

# Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# File Handling
aiofiles==23.2.1
python-magic==0.4.27

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# CORS
python-multipart==0.0.6

# Additional security packages for messaging
fernet==1.0.1
